requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
requests>=2.32.5,<3.0.0
pytz>=2025.2,<2026.0
PyYAML>=6.0.3,<7.0.0
feedparser>=6.0.0,<7.0.0
boto3>=1.35.0,<2.0.0
